sync master with development

+++ test/e2e.utils.js
@@ -1,15 +1,20 @@
 'use strict';
 
+const _ = require('lodash');
 const shell = require('shelljs');
 const e2eEnv = require('./e2e.env');
 const wsApi = require('supertest')(e2eEnv.wsUrl);
+const expect = require('chai').expect;
+
+const START_WAFFLE_SERVER = (process.env.START_WAFFLE_SERVER !== 'false');
+const DROP_MONGO_DATABASE = (process.env.DROP_MONGO_DATABASE !== 'false');
 
 module.exports = {
   dropMongoDb,
   stopWaffleServer,
   startWaffleServer,
   setUpEnvironmentVariables,
-  sendDdfqlRequest
+  sendDdfqlRequestAndVerifyResponse
 };
 
 function sendDdfqlRequest(ddfql, onResponseReceived) {
@@ -23,15 +28,21 @@ function sendDdfqlRequest(ddfql, onResponseReceived) {
 
 function startWaffleServer() {
   setUpEnvironmentVariables();
-  shell.exec(`./node_modules/.bin/forever start -t -o /dev/null -l /dev/null -a --uid "${e2eEnv.wsUid}" server.js`);
+  if (START_WAFFLE_SERVER) {
+    shell.exec(`./node_modules/.bin/forever start -t -o /dev/null -l /dev/null -a --uid "${e2eEnv.wsUid}" server.js`);
+  }
 }
 
 function stopWaffleServer() {
-  shell.exec(`./node_modules/.bin/forever stop "${e2eEnv.wsUid}"`);
+  if (START_WAFFLE_SERVER) {
+    shell.exec(`./node_modules/.bin/forever stop "${e2eEnv.wsUid}"`);
+  }
 }
 
 function dropMongoDb() {
-  shell.exec(`mongo ${e2eEnv.mongodb} --eval "db.dropDatabase()"`);
+  if (DROP_MONGO_DATABASE) {
+    shell.exec(`mongo ${e2eEnv.mongodb} --eval "db.dropDatabase()"`);
+  }
 }
 
 function setUpEnvironmentVariables() {
@@ -40,4 +51,16 @@ function setUpEnvironmentVariables() {
   shell.env['NODE_ENV'] = e2eEnv.nodeEnv;
   shell.env['DEFAULT_USER_PASSWORD'] = e2eEnv.pass;
   shell.env['INNER_PORT'] = e2eEnv.wsPort;
+}
+
+function sendDdfqlRequestAndVerifyResponse(ddfql, expectedResponse, done) {
+  sendDdfqlRequest(ddfql, (error, response) => {
+    const actualRows = _.sortBy(response.body.rows);
+    const expectedRows = _.sortBy(expectedResponse.rows);
+
+    expect(actualRows).to.deep.equal(expectedRows);
+    expect(response.body.headers).to.deep.equal(expectedResponse.headers);
+
+    done();
+  });
 }
